Salvage macro for writing docstrings for unit tests
I wrote this a while ago. It makes it easy to write really lame
docstrings for unit tests in Python. Current employer mandates PEP8,
even for unit tests, so docstrings are required. The above macro will
allow you to position your cursor anywhere on the method name of a unit
test, and creates a docstring based on that method name. So, e.g., this:

    def test_frobnicator_frobnicates_properly(self):

into this:

    def test_frobnicator_frobnicates_properly(self):
        """Test frobnicator frobnicates properly."""