Correctly set window height and width on Linux
On my ILM machine running CentOS 6 and KDE4, the window height and width
was not set correctly unless the values were explicitly set (instead of
using `set lines=999`, for example). Instead, the Gvim window took on
the same dimensions as the Konsole window from which it was launched.
Not sure why this was the case, but this is a hack to make it work for
now.