Install vim-operator-surround plugin
I think this is a bit easier to use than Tim Pope's surround plugin. I
changed the suggested default keybindings to use S instead of s, though,
since I really like the functionality of s (but don't like the delay
caused by the new operators). Of course, it makes S (substitute an
entire line) harder to use now, but S is basically the same as ^C (that
is, caret + c, not control-c) anyway, and I rarely use it (if at all) so
I'll give this mapping a try for now.

More info at [1].

  [1] http://www.reddit.com/r/vim/comments/2qozgi