Only set GTK2-specific settings if the GUI is also running
This fixes an issue that I noticed when I compiled a custom version of
vim 7.4 on my ILM machine; the terminal was all messed up when I exited
(terminal) vim. I tracked it down to the fact that the GTK2 settings
were running, which changed the window geometry, which messed up either
bash or Konsole. These settings should only be set in the GUI, so only
do them if the GUI is running.