Create lightline color theme for base16-ocean
It's not really perfect, and it's a little bland, but it's a start and I
suppose I can tweak it as I go along. The colors come from [1], so I can
change them to suit my needs as I get adventurous.

The theme was copied from the 16color theme, bundled with lightline, and
modified with colors from the base16-ocean palette. I honestly don't
know what the second number (the base-10 integer) in each bracket pair
actually means, so I didn't bother to change that. <shrug> Whatever, it
works well enough for now.