Use filebeagle instead of vinegar
Vinegar has this annoying habit of leaving the file listing buffer open,
which pollutes Vim's buffer list. There is a debate as to whether it is
vinegar's or netrw's fault [1], but regardless, it's a problem that
still hasn't been fixed. filebeagle does not have this issue and
operates in nearly the same way as vinegar, so I'm going to start using
that instead.

  [1] https://github.com/tpope/vim-vinegar/issues/13