Use full path to `screen-size` binary
When launched from the Finder or the Dock (e.g., by double-clicking the
application icon or double-clicking an associated file), MacVim does not
have access to my configured $PATH variable and does not have
~/.local/bin on its $PATH, so it cannot find the `screen-size` binary,
and thus the window size does not get set properly, so we must hardcode
the path to the binary in the gui.vim config.

I don't like this solution in theory because it means I always have to
install the `screen-size` binary at ~/.local/bin, but in practice I do
anyway, so I'll wait to come up with a more complicated until I
absolutely have to.