Use native Vim 8 package manager instead of Vundle
Vundle was a bit outdated and lacked features, such as the ability to
checkout a branch of a plugin, which is required for a plugin I want to
install. There are other package managers, but I am now using Vim 8 on
every computer I use, and since Vim 8 came out over 8 years ago, I
imagine just about every computer I use will probably have Vim 8 or
newer -- which is to say that it should be fine to use its native
package manager instead of a finding a third-party manager. It's nice to
be able to use git submodules to check out every plugin on clone, too.

I will miss running `PluginInstall!` to update all my plugins in one
swell foop, but I can probably write a function to do that for me.