Install Coc.nvim to highlight TypeScript bullshit
I had to give in and make Vim into an IDE, but there's virtually no
other way to work efficiently with TypeScript. Thanks to [1], [2], and
[3], I was able to get CoC more or less working.

There are some annoyances, like the auto-complete that pops up _all the
damn time_, but hopefully I can fix it to only pop up in TypeScript
files.

  [1]: https://pragmaticpineapple.com/ultimate-vim-typescript-setup/
  [2]: https://pragmaticpineapple.com/ultimate-vim-vue-setup/
  [3]: https://thoughtbot.com/blog/modern-typescript-and-react-development-in-vim