Revert "Use native Vim 8 package manager instead of Vundle"
The native package manager is...okay. I do like that I could easily
navigate to the source of plugins when viewing my repository in GitHub.
But submodules are a pain to manage manually, especially when you have a
lot of them, and especially when you want to remove one. Vundle provides
some much nicer tooling to update plugins, cleanup plugins that have
been removed, and automatically generate help tags.

And the original impetus for the the migration is no longer valid, as I
don't plan to use CoC.nvim right now anyway, so the fact that Vundle
doesn't allow me to check out a particular branch isn't a huge problem.
Maybe I will have to switch to another newfangled package manager in the
future, or maybe if I do decide to use CoC.nvim, I'll have to see if I
can use both the native package manager and Vundle...but we'll cross
that bridge when we come to it.

To revert an existing checkout that uses the native package manager:

    $ cd ~/.vimfiles
    $ git pull
    $ rm -rf ~/.vim/pack
    $ vim .git/config
    # And remove all the submodule junk
    $ rm -rf ~/.git/modules
    $ vim
    # And run :PluginInstall!
    # Then exit and run vim again to make sure everything works
    $ vim

This reverts commit 5c885493bc3ddc8f7631170b15341fe6622cc51b.