Fix toggling of window sizes on larger screens
I now have a screen that is 2560 pixels in width. By default,
MacVim/gVim correctly opened a window of 192 pixels on this screen.
However, `ToggleWindowSize()` incorrectly assumed that 192 pixels was
the maximum width, so trying to toggle the window size to make the
window bigger would actually shrink the window to 96 pixels. This now
fixes `ToggleWindowSize()` so it _expands_ on larger screens.
Furthermore, the expansion is based on the size of the shrunken window:
2x on smaller screens and 1.5x on larger screens.

I also moved `ToggleWindowSize()` so it is available to both MacVim and
gVim, mostly because `ToggleWindowSize()` now accesses script-local
variables.

I also store the screen width in a script-local variable because calling
`screen-size` in every invocation of the `ToggleWindowSize()` function
(or, actually, calling it twice) resulting in a noticeable delay in
screen resizing.